refactor: TextField에서 VisualTransformation을 내부 객체로 사용하도록 변경
