+++ app/src/main/java/woowacourse/payments/ui/newcard/textfields/CardNumberTextField.kt
@@ -11,12 +11,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.OffsetMapping
+import androidx.compose.ui.text.input.TransformedText
+import androidx.compose.ui.text.input.VisualTransformation
 import woowacourse.payments.R
 import woowacourse.payments.domain.CardNumber
 import woowacourse.payments.ui.theme.Gray
-
-private const val CARD_NUMBER_REQUIRED_LENGTH = 16
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
@@ -43,7 +45,7 @@ fun CardNumberTextField(
         value = text.value,
         onValueChange = { newValue: String -> updateValue(newValue) },
         singleLine = true,
-        visualTransformation = CardNumberTransformation,
+        visualTransformation = visualTransformation,
         label = { Text(stringResource(R.string.card_number_label)) },
         placeholder = {
             Text(
@@ -65,3 +67,34 @@ fun CardNumberTextField(
         keyboardActions = KeyboardActions(onDone = { focusManager.moveFocus(FocusDirection.Next) }),
     )
 }
+
+private const val CARD_NUMBER_REQUIRED_LENGTH = 16
+private const val CARD_NUMBER_CHUNK_SIZE = 4
+private const val CARD_NUMBER_DELIMITER = " - "
+
+private val visualTransformation: VisualTransformation =
+    object : VisualTransformation {
+        override fun filter(text: AnnotatedString): TransformedText =
+            TransformedText(
+                AnnotatedString(
+                    text.text.chunked(CARD_NUMBER_CHUNK_SIZE).joinToString(CARD_NUMBER_DELIMITER),
+                ),
+                offsetMapping,
+            )
+    }
+
+private val offsetMapping =
+    object : OffsetMapping {
+        override fun originalToTransformed(offset: Int): Int {
+            val multiplier = (offset - 1).coerceAtLeast(0) / CARD_NUMBER_CHUNK_SIZE
+            return offset + CARD_NUMBER_DELIMITER.length * multiplier
+        }
+
+        override fun transformedToOriginal(offset: Int): Int {
+            val multiplier =
+                ((offset - 1)).coerceAtLeast(0) / (CARD_NUMBER_CHUNK_SIZE + CARD_NUMBER_DELIMITER.length)
+            return (offset - (CARD_NUMBER_DELIMITER.length * multiplier)).coerceAtMost(
+                CARD_NUMBER_CHUNK_SIZE * (multiplier + 1),
+            )
+        }
+    }
